+++ packages/@expo/cli/src/run/ios/runIosAsync.ts
@@ -22,7 +22,7 @@ import { getValidBinaryPathAsync } from './validateExternalBinary';
 import { exportEagerAsync } from '../../export/embed/exportEager';
 import { getContainerPathAsync, simctlAsync } from '../../start/platforms/ios/simctl';
 import { CommandError } from '../../utils/errors';
-import { resolveRemoteBuildCache } from '../resolveRemoteBuildCache';
+import { resolveRemoteBuildCache } from '../remoteBuildCache';
 
 const debug = require('debug')('expo:run:ios');
